Upgrades specs2 version Fixes Lint errors on console Moves to travis
