+++ .travis.yml
@@ -1,0 +1,20 @@
+language: scala
+scala:
+  - 2.11.6
+before_install:
+  - sudo apt-get update -qq
+  - sudo apt-get install libc6-i386 lib32z1 lib32stdc++6
+  - wget http://dl.google.com/android/android-sdk_r24-linux.tgz
+  - tar xf android-sdk_r24-linux.tgz
+  - export ANDROID_HOME=$PWD/android-sdk-linux
+  - export ANDROID_SDK_HOME=$PWD/android-sdk-linux
+  - export PATH=${PATH}:${ANDROID_HOME}/tools:${ANDROID_HOME}/platform-tools
+  - echo yes | android update sdk --all --filter platform-tools --no-ui
+  - echo yes | android update sdk --all --filter build-tools-22.0.1 --no-ui
+  - echo yes | android update sdk --all --filter android-22 --no-ui
+  - echo yes | android update sdk --all --filter extra-android-support --no-ui
+  - echo yes | android update sdk --all --filter extra-android-m2repository --no-ui
+  - echo yes | android update sdk --all --filter extra-google-m2repository --no-ui
+  - echo yes | android update sdk --all --filter extra-google-google_play_services --no-ui
+script:
+  - sbt ++$TRAVIS_SCALA_VERSION clean test
