--- circle.yml
@@ -1,7 +1,0 @@
-machine:
-  environment:
-    ANDROID_HOME: /usr/local/android-sdk-linux
-
-dependencies:
-  pre:
-    - echo y | android update sdk --no-ui --all --filter "build-tools-22.0.1,extra-android-m2repository,extra-android-support,extra-google-m2repository,extra-google-google_play_services"
