+++ project/plugins.sbt
@@ -1,1 +1,1 @@
-addSbtPlugin("com.hanhuy.sbt" % "android-sdk-plugin" % "1.3.22")
+addSbtPlugin("com.hanhuy.sbt" % "android-sdk-plugin" % "1.4.2")
