+++ src/test/scala/com/fortysevendeg/scala/android/ContextWrapperContextTestSupport.scala
@@ -1,13 +1,13 @@
 package com.fortysevendeg.scala.android
 
+import android.content.Context
 import macroid.ContextWrapper
 import org.specs2.mock.Mockito
-import org.specs2.specification.Scope
 
-trait ContextWrapperTestSupport
-    extends Mockito
-    with TestConfig
-    with Scope {
+trait ContextWrapperContextTestSupport
+    extends Mockito {
+
+  val mockContext = mock[Context]
 
   implicit val context: ContextWrapper = mock[ContextWrapper]
