+++ src/test/scala/com/fortysevendeg/scala/android/modules/forecast/ForecastServicesSpec.scala
@@ -2,29 +2,33 @@ package com.fortysevendeg.scala.android.modules.forecast
 
 import com.fortysevendeg.scala.android.modules.forecast.impl.ForecastServices
 import com.fortysevendeg.scala.android.modules.forecast.model._
-import com.fortysevendeg.scala.android.{ContextWrapperTestSupport, BaseTestSupport}
+import com.fortysevendeg.scala.android.{ContextWrapperContextTestSupport, BaseTestSpecification}
 import com.squareup.okhttp.OkHttpClient
 import io.taig.communicator.result.Parser
 import macroid.ContextWrapper
-import org.specs2.mutable.Specification
 
 import scala.concurrent.duration.Duration
 import scala.concurrent.{Await, Future}
 
-trait ForecastServiceMock
-  extends ForecastServices
-  with ContextWrapperTestSupport {
+trait ForecastServiceSpecification
+  extends BaseTestSpecification
+  with ContextWrapperContextTestSupport {
 
-  override def loadJsonUrl(latitude: Double, longitude: Double)(implicit context: ContextWrapper): String = "http://fake_url/"
+  trait ForecastServiceScope
+    extends BaseTestScope
+    with ForecastServices
+    with ForecastServiceData {
 
-  override def loadHeaderTuple(implicit context: ContextWrapper): (String, String) = ("key_name", "key_value")
+    override def loadJsonUrl(latitude: Double, longitude: Double)(implicit context: ContextWrapper): String = "http://fake_url/"
+
+    override def loadHeaderTuple(implicit context: ContextWrapper): (String, String) = ("key_name", "key_value")
+
+  }
 
 }
 
-class ForecastServicesSpec
-  extends Specification
-  with BaseTestSupport {
-  
+trait ForecastServiceData {
+
   val validJson: ApiModel = ApiModel(
     id = 5809844,
     dt = 1425060066,
@@ -33,31 +37,35 @@ class ForecastServicesSpec
     name = "Seattle",
     coord = ApiCoord(-122.38, 47.66),
     sys = ApiSys(
-      message = Some(0.0168), 
-      country = Some("US"), 
-      sunrise = Some(1425048757), 
+      message = Some(0.0168),
+      country = Some("US"),
+      sunrise = Some(1425048757),
       sunset = Some(1425088297)),
     main = ApiMain(
-      temp = Some(6.636), 
-      temp_min = Some(6.636), 
-      temp_max = Some(6.636), 
-      pressure = Some(1009.8), 
-      sea_level = Some(1021.71), 
-      grnd_level = Some(1009.8), 
+      temp = Some(6.636),
+      temp_min = Some(6.636),
+      temp_max = Some(6.636),
+      pressure = Some(1009.8),
+      sea_level = Some(1021.71),
+      grnd_level = Some(1009.8),
       humidity = Some(100)),
     weather = Seq(ApiWeather(id = 501, main = "Rain", description = "moderate rain", icon = "10d")),
     wind = ApiWind(speed = Some(2.17), deg = Some(36.0011)),
     rain = Some(Map("3h" -> 3.5)),
     clouds = ApiClouds(all = Some(92)))
-  
+
+}
+
+class ForecastServicesSpec extends ForecastServiceSpecification {
+
   "ForecastServices component" should {
-    
+
     "return forecast with right JSON" in
-      new ForecastServiceMock {
+      new ForecastServiceScope {
 
         override def loadJson[T](url: String, headers: Seq[(String, String)])(implicit parser: Parser[T], client: OkHttpClient = new OkHttpClient()): Future[T] =
           Future.successful[T](validJson.asInstanceOf[T])
-        
+
         val forecast = Forecast(
           Location(5809844, "Seattle", 47.66, -122.38),
           Some(Weather(501, "Rain", "moderate rain", "10d", Some(6.636))))
@@ -66,7 +74,7 @@ class ForecastServicesSpec
       }
 
     "got the exception thrown by the call to loadJson" in
-      new ForecastServiceMock {
+      new ForecastServiceScope {
 
         override def loadJson[T](url: String, headers: Seq[(String, String)])(implicit parser: Parser[T], client: OkHttpClient = new OkHttpClient()): Future[T] =
           Future.failed[T](new RuntimeException())
